+++ Sources/InstanaAgent/Configuration/VersionConfig.swift
@@ -1,3 +1,3 @@
 struct VersionConfig {
-    static let agentVersion = "1.6.8"
+    static let agentVersion = "1.6.9"
 }